provisioning unit tests now divided

+++ provisioning/tests.py
@@ -2,8 +2,6 @@ import random
 import string
 import time
 
-from bs4 import BeautifulSoup as bs
-
 from django.apps import apps
 from django.conf import settings
 from django.contrib.auth import get_user_model
@@ -26,8 +24,9 @@ def randomString(stringLength=10):
 
 _MAX_HTML_PAGE_LEN = 6038
 
-_uid = 'unittest_john1298731289371283123-kgkgk-{}'.format(randomString())
-_passwd = 'Ssmith_67!'
+_uid = 'unit_test_user' #-{}'.format(randomString())
+_new_uid = '_new__'+ _uid
+_passwd = 'Ssmi#th_67'
 matricola_studente =  '111190{}'.format(randomString())
 matricola_dipendente = '982388{}'.format(randomString())
 #  additional_affiliations = [('otherinst.net', 'student', matricola_studente),
@@ -51,18 +50,8 @@ class ProvisioningTestCase(TestCase):
     databases = '__all__'
 
     def setUp(self):
-        """test identity creation"""
-        # cleanup
-        pass
-
-    def test_entire_provisioning(self):
-        """test identity account provisioning"""
-        # disable username with prefix constraint
-        settings.ACCOUNT_CREATE_USERNAME_PRESET = None
-
-        d = LdapAcademiaUser.objects.filter(uid__contains='unittest_john1298731289371283123-kgkgk')
-        if d and _PURGE_LDAP_TEST_USER:
-            d.delete()
+        # clean up
+        self.tearDown()
 
         d = Identity.objects.create(**_test_guy)
         # additional affiliations
@@ -72,37 +61,52 @@ class ProvisioningTestCase(TestCase):
                                                  name=addaff[1],
                                                  unique_code=addaff[2])
 
-        d = Identity.objects.get(mail=_test_guy['mail'])
-        p = IdentityProvisioning.objects.create(identity=d)
-        token_url = p.get_activation_url()
+        self.identity = Identity.objects.get(mail=_test_guy['mail'])
+        self.provisioning = IdentityProvisioning.objects.create(identity=self.identity)
+
+        # provisioning test
+        token_url = self.provisioning.get_activation_url()
         c = Client()
+        # test identity creation
+        # disable username with prefix constraint
+        settings.ACCOUNT_CREATE_USERNAME_PRESET = None
         request = c.post(token_url, {'username': _uid,
                                      'password': _passwd,
                                      'password_verifica': _passwd,
                                      'mail': _test_guy['mail']})
-        # check = (b'errorlist' in request.content)
-        check = (b'alert-error' in request.content)
-        if check:
-            print(bs(request.content, features="html.parser").prettify()[:_MAX_HTML_PAGE_LEN])
+        check = (b'error' in request.content)
         self.assertIs(check, False)
 
         # check real user existence
-        d = LdapAcademiaUser.objects.filter(uid=_uid).first()
-        self.assertTrue(d)
-        self.assertTrue(d.userPassword)
-        self.assertTrue(d.sambaNTPassword)
-        # print(d.userPassword)
+        self.ldap_user = LdapAcademiaUser.objects.filter(uid=_uid).first()
+        self.assertTrue(self.ldap_user)
+        self.assertTrue(self.ldap_user.userPassword)
+        self.assertTrue(self.ldap_user.sambaNTPassword)
 
-    # def test_b_login(self):
+        # first login
+        c = Client()
+        lurl = reverse('provisioning:provisioning_login')
+        request = c.post(lurl, {'username': _uid,
+                                'password': _passwd})
+        self.assertEqual(request.status_code, 302)
+
+
+    def test_print_ldap_user_Attributes(self):
+        print('\nCreated User Attributes:\n')
+        for k,v in self.ldap_user.__dict__.items():
+            if k[0] == '_': continue
+            print('\t{}: {}'.format(k, v))
+
+
+    def test_b_login(self):
         """test account login"""
         c = Client()
         lurl = reverse('provisioning:provisioning_login')
         request = c.post(lurl, {'username': _uid,
                                 'password': _passwd})
-        # self.assertEqual(int(request.status_code), 302)
         self.assertEqual(request.status_code, 302)
 
-    # def test_c_password_change(self):
+    def test_c_password_change(self):
         """test a password change with confirmation"""
         c = Client()
         # login first
@@ -119,22 +123,18 @@ class ProvisioningTestCase(TestCase):
         check = (b'alert-error' in request.content)
         self.assertFalse(check)
 
-    # def test_d_changedata(self):
+    def test_d_changedata(self):
         """test account change identity data"""
         c = Client()
         lurl = reverse('provisioning:provisioning_login')
-        request = c.post(lurl, {'username': _uid,
-                                'password': _passwd+_passwd})
+        request = c.post(lurl, {'username': _uid, 'password': _passwd})
         self.assertEqual(request.status_code, 302)
 
         lurl = reverse('provisioning:change_data')
         request = c.post(lurl, {'mail': 'ingo_'+_test_guy['mail'],
                                 'telephoneNumber': '0984567683'})
         self.assertEqual(request.status_code, 200)
-        #self.assertEqual(request.status_code, 302)
         check = (b'errorlist' in request.content)
-        # if check:
-            # print(bs(request.content, features="html.parser").prettify()[:_MAX_HTML_PAGE_LEN])
         self.assertIs(check, False)
 
         # check change token
@@ -163,7 +163,7 @@ class ProvisioningTestCase(TestCase):
         if _WAIT_FOR_A_CHECK:
             time.sleep(6000)
 
-    # def test_change_username(self):
+    def test_change_username(self):
         user_model = apps.get_model(settings.AUTH_USER_MODEL)
         django_user = user_model.objects.get(username=_uid)
         self.assertFalse(django_user.change_username)
@@ -172,7 +172,7 @@ class ProvisioningTestCase(TestCase):
         # login
         lurl = reverse('provisioning:provisioning_login')
         request = c.post(lurl, {'username': _uid,
-                                'password': _passwd+_passwd})
+                                'password': _passwd})
         self.assertEqual(request.status_code, 302)
 
         # try to access to change username page (fails!)
@@ -207,7 +207,6 @@ class ProvisioningTestCase(TestCase):
 
         # choose a valid new username
         lurl = reverse('provisioning:change_username')
-        _new_uid = '_new__'+ _uid
         request = c.post(lurl, {'old_username': _uid,
                                 'new_username': _new_uid,
                                 'confirm_new_username': _new_uid},
@@ -233,7 +232,7 @@ class ProvisioningTestCase(TestCase):
         # login
         lurl = reverse('provisioning:provisioning_login')
         request = c.post(lurl, {'username': _new_uid,
-                                'password': _passwd+_passwd})
+                                'password': _passwd})
         self.assertEqual(request.status_code, 302)
 
         # try to access to change username page
@@ -242,12 +241,12 @@ class ProvisioningTestCase(TestCase):
         request = c.get(change_username_url, follow=True)
         self.assertEqual(request.status_code, 403)
 
-    # def test_d_clean(self):
-        # """cleanup"""
-        d = LdapAcademiaUser.objects.filter(uid=_new_uid).first()
-        print('\nCreated User Attributes:\n')
-        for k,v in d.__dict__.items():
-            if k[0] == '_': continue
-            print('\t{}: {}'.format(k, v))
-        if _PURGE_LDAP_TEST_USER:
+        # come back to the previous one
+        lu.uid=_uid
+        lu.save()
+
+    def tearDown(self):
+        """cleanup"""
+        d = LdapAcademiaUser.objects.filter(uid=_uid).first()
+        if _PURGE_LDAP_TEST_USER and d:
             d.delete()
